cli/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "cc"
version = "0.1.0"
description = "cc CLI bootstrap"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "typer",
    "requests",
    "python-dotenv"
]

[project.scripts]
cc = "cc.cli:app"

[tool.setuptools.packages.find]
where = ["."]
